Use ssize_t over size_t to silence warnings

+++ src/temperature_threshold_manager.c
@@ -1,7 +1,5 @@
 ﻿#include "temperature_threshold_manager.h"
 
-#include <stddef.h> // size_t
-
 bool TemperatureThresholdManager_LegacyBehaviour = false;
 
 Error* ThresholdManager_Init(ThresholdManager* self, array_of(TemperatureThreshold)* thresholds) {
@@ -9,8 +7,8 @@ Error* ThresholdManager_Init(ThresholdManager* self, array_of(TemperatureThresho
   my.thresholds = *thresholds;
 
   /* Bubble sort - ascending */
-  for (size_t i = 0; i < thresholds->size - 1; ++i)
-    for (size_t j = 0; j < thresholds->size - i - 1; ++j) {
+  for (ssize_t i = 0; i < thresholds->size - 1; ++i)
+    for (ssize_t j = 0; j < thresholds->size - i - 1; ++j) {
       TemperatureThreshold* a = &thresholds->data[j];
       TemperatureThreshold* b = &thresholds->data[j+1];
       if (a->UpThreshold > b->UpThreshold) {
@@ -24,11 +22,11 @@ Error* ThresholdManager_Init(ThresholdManager* self, array_of(TemperatureThresho
 }
 
 TemperatureThreshold* ThresholdManager_AutoSelectThreshold(ThresholdManager* self, float temperature) {
-  const int size = my.thresholds.size;
+  const ssize_t size = my.thresholds.size;
   if (! size)
     return NULL;
 
-  int i = my.current ? my.current - my.thresholds.data : 0;
+  ssize_t i = my.current ? my.current - my.thresholds.data : 0;
   if (TemperatureThresholdManager_LegacyBehaviour) {
     while (i > 0        && temperature <= my.thresholds.data[i].DownThreshold)   --i;
     while (i < size - 1 && temperature >= my.thresholds.data[i + 1].UpThreshold) ++i;
